Revert "Merge branch 'dev' of https://github.com/ubclaunchpad/forum into fe-courses-v2"
This reverts commit 4626fba4208c3f323fc578a67b9409b9a5827596, reversing
changes made to d99e998059db443e5b9a79b7ab870a0bfefb6371.
