+++ shared/schema/course.ts
@@ -71,8 +71,6 @@ export const defaultPermissionsSchema = z.object({
   }),
 });
 
-export type Permissions = z.infer<typeof defaultPermissionsSchema>;
-
 export const AccountRolesSchema = z.object({
   id: z.string(),
   name: z.string(),
@@ -88,7 +86,6 @@ export type DefaultRoles = (typeof DEFAULT_ROLES)[number];
 export const instructorRole: DefaultRoles = DEFAULT_ROLES[0];
 export const staffRole: DefaultRoles = DEFAULT_ROLES[1];
 export const studentRole: DefaultRoles = DEFAULT_ROLES[2];
-export const defaultRoleSchema = z.enum(DEFAULT_ROLES);
 
 export const CourseRolesSchema = z.object({
   course_id: z.string(),
